# [Don't merge] Proof of concept for caching the linkables endpoint