This is a quick demonstration of how we could speed up the linkables endpoint, which has been timing out frequently on applications such as content-tagger.

It works by caching the result of the linkables call. To make the cache self-refreshing, we use a cache key composed of the `document_type` and the last updated time of a the document_type. For example, the cache key generated on my machine was `linkables/policy/2016-05-17+14%3A42%3A52+UTC`.

Which means that if a `policy` is updated, `updated_at` will change and a cache-miss will occur and the content will be refreshed. We would then depend on memcached LRU mechanism to evict old content.

This is called key-based cached expiration and is pretty cool. 

https://signalvnoise.com/posts/3113-how-key-based-cache-expiration-works

Some silly benchmarking below (done in development, with `FileStore`, so the end result could actually be faster).

### Before

```
@tijmenb ~/govuk/publishing-api on master $ ab -n 100 -c 1 http://publishing-api.dev.gov.uk/v2/linkables?document_type=topic
This is ApacheBench, Version 2.3 <$Revision: 1663405 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking publishing-api.dev.gov.uk (be patient).....done


Server Software:        nginx
Server Hostname:        publishing-api.dev.gov.uk
Server Port:            80

Document Path:          /v2/linkables?document_type=topic
Document Length:        96354 bytes

Concurrency Level:      1
Time taken for tests:   17.435 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      9719400 bytes
HTML transferred:       9635400 bytes
Requests per second:    5.74 [#/sec] (mean)
Time per request:       174.350 [ms] (mean)
Time per request:       174.350 [ms] (mean, across all concurrent requests)
Transfer rate:          544.40 [Kbytes/sec] received

Connection Times (ms)
             min  mean[+/-sd] median   max
Connect:        0   25 144.4      0    1004
Processing:   127  149  10.6    150     183
Waiting:      126  147  10.5    148     181
Total:        127  174 145.1    150    1150

Percentage of the requests served within a certain time (ms)
 50%    150
 66%    156
 75%    157
 80%    158
 90%    165
 95%    178
 98%    949
 99%   1150
100%   1150 (longest request)
@tijmenb ~/govuk/publishing-api on master $ 
```


### After

```
@tijmenb ~/govuk/publishing-api on poc-caching $ ab -n 100 -c 1 http://publishing-api.dev.gov.uk/v2/linkables?document_type=topic
This is ApacheBench, Version 2.3 <$Revision: 1663405 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking publishing-api.dev.gov.uk (be patient).....done


Server Software:        nginx
Server Hostname:        publishing-api.dev.gov.uk
Server Port:            80

Document Path:          /v2/linkables?document_type=topic
Document Length:        96354 bytes

Concurrency Level:      1
Time taken for tests:   3.881 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      9719400 bytes
HTML transferred:       9635400 bytes

Requests per second:    25.77 [#/sec] (mean)
Time per request:       38.806 [ms] (mean)
Time per request:       38.806 [ms] (mean, across all concurrent requests)
Transfer rate:          2445.90 [Kbytes/sec] received

Connection Times (ms)
             min  mean[+/-sd] median   max
Connect:        0   10 100.3      0    1003
Processing:    24   28   5.4     27      58
Waiting:       21   26   5.5     25      56
Total:         24   39 100.9     27    1036

Percentage of the requests served within a certain time (ms)
 50%     27
 66%     28
 75%     30
 80%     31
 90%     36
 95%     40
 98%     58
 99%   1036
100%   1036 (longest request)
@tijmenb ~/govuk/publishing-api on poc-caching $
```