@benhyland 2016-05-20
How does this interact with the denormalisation work Eliot is doing in Publishing Platform? Are both pieces of work necessary?