@tijmenb 2016-05-20
I think it's probably best to wait for @elliotcm's work on derived presentations. I'm hoping that that improves linkables performance enough so that we don't need the complexity of an extra caching layer.